+++ packages/core/src/components/BreadcrumbsBar/BreadcrumbItem/BreadcrumbContent/BreadcrumbContent.tsx
@@ -1,9 +1,10 @@
 /* eslint-disable jsx-a11y/no-noninteractive-tabindex */
 import React, { ForwardedRef, forwardRef, useCallback, useMemo } from "react";
 import { keyCodes } from "../../../../constants";
-import { SubIcon, VibeComponentProps } from "../../../../types";
+import { VibeComponentProps } from "../../../../types";
 import cx from "classnames";
 import styles from "./BreadcrumbContent.module.scss";
+import { SubIcon } from "../../../Icon";
 
 export interface BreadcrumbContentProps extends VibeComponentProps {
   isClickable?: boolean;
