+++ packages/core/src/components/ColorPicker/ColorPicker.types.ts
@@ -1,3 +1,5 @@
 export type ColorShapes = "square" | "circle";
 
 export type ColorPickerSizes = "small" | "medium" | "large";
+
+export type ColorStyle = "regular" | "hover" | "selected";
