+++ packages/core/src/components/Combobox/components/ComboboxConstants.ts
@@ -1,6 +1,6 @@
-import { SubIcon } from "../../../types";
 import React from "react";
 import { MutableRef } from "preact/hooks";
+import { SubIcon } from "../../Icon";
 
 export const COMBOBOX_DIVIDER_ITEM = "combobox-divider";
 export const COMBOBOX_CATEGORY_ITEM = "combobox-category";
