+++ packages/core/src/components/Menu/MenuItem/MenuItem.tsx
@@ -3,8 +3,8 @@ import React, { AriaAttributes, ForwardedRef, ReactElement, forwardRef, useMemo,
 import Tooltip, { TooltipProps } from "../../../components/Tooltip/Tooltip";
 import Icon from "../../../components/Icon/Icon";
 import useIsOverflowing from "../../../hooks/useIsOverflowing/useIsOverflowing";
-import { SubIcon, VibeComponent, VibeComponentProps, withStaticProps } from "../../../types";
-import { IconType } from "../../Icon";
+import { VibeComponent, VibeComponentProps, withStaticProps } from "../../../types";
+import { IconType, SubIcon } from "../../Icon";
 import { CloseMenuOption, MenuChild } from "../Menu/MenuConstants";
 import Label from "../../Label/Label";
 import styles from "./MenuItem.module.scss";
