+++ packages/core/src/components/Toggle/Toggle.tsx
@@ -6,7 +6,7 @@ import { MockToggle } from "./MockToggle";
 import { VibeComponent, VibeComponentProps } from "../../types";
 import styles from "./Toggle.module.scss";
 import { ComponentDefaultTestId, getTestId } from "../../tests/test-ids-utils";
-import { Sizes } from "./Toggle.types";
+import { ToggleSizes } from "./Toggle.types";
 
 export interface ToggleProps extends VibeComponentProps {
   /**
@@ -24,7 +24,7 @@ export interface ToggleProps extends VibeComponentProps {
   offOverrideText?: string;
   ariaLabel?: string;
   ariaControls?: string;
-  size?: Sizes;
+  size?: ToggleSizes;
 }
 
 const Toggle: VibeComponent<ToggleProps, HTMLInputElement> = forwardRef(
