+++ packages/core/src/components/Tooltip/Tooltip.types.ts
@@ -1,3 +1,8 @@
 export type TooltipPositions = "top" | "right" | "bottom" | "left";
 
 export type TooltipTheme = "dark" | "primary";
+
+export type MoveBy = {
+  main?: number;
+  secondary?: number;
+};
