--- packages/core/src/types/ArrayLastElement.ts
@@ -1,1 +1,0 @@
-export type ArrayLastElement<T extends unknown[]> = T extends [...unknown[], infer R] ? R : never;
