--- packages/core/src/types/MoveBy.ts
@@ -1,4 +1,0 @@
-export type MoveBy = {
-  main?: number;
-  secondary?: number;
-};
