--- packages/core/src/types/SubIcon.ts
@@ -1,4 +1,0 @@
-import React from "react";
-import { IconSubComponentProps } from "src/components/Icon/Icon";
-
-export type SubIcon = string | React.FC<IconSubComponentProps> | null;
