+++ packages/core/src/types/index.ts
@@ -1,10 +1,6 @@
 export * from "./events";
-export * from "./SubIcon";
 export { default as VibeComponent } from "./VibeComponent";
 export { default as VibeComponentProps } from "./VibeComponentProps";
 export * from "./withStaticProps";
-export * from "./ArrayLastElement";
-export * from "./SplitString";
 export * from "./ElementContent";
-export * from "./ValueOf";
-export * from "./Colors";
+export * from "../components/Chips/Colors.types";
